requirements.txt
streamlit
opencv-python-headless
Pillow
openai
numpy
